+++ src/components/delete-modal/delete-modal.css
@@ -36,7 +36,7 @@
 }
 
 .delete_form_buttons {
-  background: url('../add-movie-form/add-form/img/CloseButton.png');
+  background: url('../film-movie-form/movie-form/img/CloseButton.png');
   width: 21px;
   height: 22px;
   padding: 0;
