+++ src/components/film-movie-form/success-modal/success-modal.css
@@ -39,7 +39,7 @@
 }
 
 .add_form_buttons {
-  background: url('../add-form/img/CloseButton.png');
+  background: url('../movie-form/img/CloseButton.png');
   width: 21px;
   height: 22px;
   padding: 0;
