+++ src/components/genres-menu/index.js
@@ -1,5 +1,5 @@
 import './genres-menu.css';
-import SortPanel from '../sort-panel';
+import { SortPanel } from '../index';
 
 const GenresMenu = () => (
   <div className='genres_menu'>
