+++ src/components/index.js
@@ -10,8 +10,22 @@ import MainPage from './main-page';
 import NumberOfFilms from './number-of-films';
 import SearchPanel from './search-panel';
 import SortPanel from './sort-panel';
+import FilmCardInformation from './film-card-information';
+import ErrorBoundary from './error-boundary';
+import DeleteModal from './delete-modal';
+import MovieForm from './film-movie-form/movie-form';
+import FilmFormModule from './film-movie-form/film-form-module';
+import FilmFormInputList from './film-movie-form/film-form-input-list';
+import SuccessModal from './film-movie-form/success-modal';
 
 export {
+  FilmFormInputList,
+  FilmCardInformation,
+  ErrorBoundary,
+  DeleteModal,
+  FilmFormModule,
+  MovieForm,
+  SuccessModal,
   Footer,
   App,
   Button,
